laga hluti

+++ properties/models.py
@@ -54,5 +54,5 @@ class ContactInfo(models.Model):
     city = models.CharField(max_length=255)
     country = models.CharField(max_length=255)
     postal_code = models.IntegerField()
-    ssn = models.IntegerField()
+    ssn = models.BigIntegerField()
